Add files via upload
In this project I worked on programming a coffee-machine simulator. The machine works with quite typical products: coffee, milk, sugar, and plastic cups; if it runs out of something, it shows a notification. The machine can make three types of coffee: espresso, cappuccino and latte. Since nothing’s for free, it also collects the money.